tboot: add libtomfastmath and libtomcrypt to the build
This adds libtomfastmath and libtomcrypt to the tboot build process
and fixes a majority of problems in the process such that tboot is
able to successfully build with the two new libraries.

Some of these changes, e.g. the X<libc_func> macros, should likely be
submitted to the upstream libtom* projects.

We should also look into replacing the existing tboot crypto/hashing
code with calls to libtomcrypt.  There is no need to carry duplicate
crypto/hash implementations.

Signed-off-by: Paul Moore <pmoore2@cisco.com>