all: add support for certificate payloads in the LCP
This commit defines a new UUID, c427f712-1043-46bf-a351-e333f67ede88,
to represent a trusted certificate payload which can be used by tboot
to perform any number of verifications.  The payload is simply a
concatenated bundle of DER encoded X509 certificates.

You can create a policy ELT with the certificate payload using a
command similar to this example:

 # lcp2_crtpolelt --create \
     --type custom --uuid certificates test.der \
     --out test.elt

Signed-off-by: Paul Moore <pmoore2@cisco.com>