all: add support for PECOFF signature verification to the tboot policy
This commit defines a new tboot policy hash type, "pecoff".  This new
type specifies that tboot should perform PECOFF signature verification
in a UEFI Secure Boot compatible manner.  If successful, a digest of
the signing authority's trust root (the trust root is defined as the
certificate in the signature verification trust chain which was
loaded from the LCP policy) will be extended into the PCR specified
in the policy.

You can create a tboot policy that specifies PECOFF signature
verification using commands similar to this example:

  # tb_polgen --create --type nonfatal \
              --alg sha256 --ctrl 0x00 pecoff.vlp
  # tb_polgen --add --num 0 --pcr 20 --hash pecoff pecoff.vlp
  # tb_polgen --show pecoff.vlp
  policy:
         version: 2
         policy_type: TB_POLTYPE_CONT_NON_FATAL
         hash_alg: TB_HALG_SHA256
         policy_control: 00000000 ()
         num_entries: 1
         policy entry[0]:
                 mod_num: 0
                 pcr: 20
                 hash_type: TB_HTYPE_PECOFF
                 num_hashes: 0

TODO:
- Document this new addition to the tboot policy, including
  recommendations about which PCRs to use
- Attempt to read the LCP even if the VLP is loaded from it's own
  nvindex (e.g. 0x01C10131 for TPM2) so we can load the trusted cert
  payload [PM: attempted in a separate patch, but doesn't appear to
  be possible with TPM2]

Signed-off-by: Paul Moore <pmoore2@cisco.com>